Several refactors for genetic.
Add in a new Context object to hold some of the particulars, and remove all state from the Engine. Combine all the mutator logic into a Generic mutator. Remove other types. Use Queryables more in the Engine to simplify flow.

Add basic caching to Random.default_uniform_random, to ensure we aren't creating a new one (and potentially seeding it the same in tight loops).